FrankenPHP standalone prepare

+++ static-build-prepare.sh
@@ -1,11 +1,19 @@
 #!/bin/bash
 
+_PWD=$(pwd)
+
 TMPDIR=./tmp
+rm -Rf $TMPDIR
 # Export the project to get rid of .git/, etc
 mkdir -p $TMPDIR/my-prepared-app1
-git archive HEAD | tar -x -C $TMPDIR/my-prepared-app1
-mv $TMPDIR/my-prepared-app1/src $TMPDIR/my-prepared-app
-rm -Rf $TMPDIR/my-prepared-app1
+if [[ $_GIT == 'yes' ]]; then
+    git archive HEAD | tar -x -C $TMPDIR/my-prepared-app1
+    mv $TMPDIR/my-prepared-app1/src $TMPDIR/my-prepared-app
+    rm -Rf $TMPDIR/my-prepared-app1
+else
+   cp -Rf src $TMPDIR/my-prepared-app
+fi
+
 cd $TMPDIR/my-prepared-app
 
 # Set proper environment variables
@@ -16,4 +24,12 @@ cp .env.prod .env
 rm -Rf tests/
 
 # Install the dependencies
-composer install --ignore-platform-reqs --no-dev -a
+podman compose run --rm -w /app/tmp/my-prepared-app php-cli \
+    composer install --ignore-platform-reqs --no-dev -a
+
+podman build -t static-app -f ${_PWD}/static-build.Dockerfile .
+podman cp $(podman create --name static-app-tmp static-app):/go/src/app/dist/frankenphp-linux-x86_64 my-app
+podman rm static-app-tmp
+mv my-app ${_PWD}/haproxy-gateway
+rm tmp -Rf
+
